Extract test blocks for `remote` and `visit` to requests spec file
Because `remote` and `visit` themselves were extracted to lib/action_creators/requests.ts, they should have their own spec file, and the action_creators spec file is large.

I moved over the minimal amount of test setup and imports to get the tests passing.

Not included: cleaning up imports to spec/lib/action_creators.spec.js which are now unused, because my IDE was not configured correctly to help identify them. They'll be cleaned up either in a later commit or in a later linting pass.
